added zoomBlock.py
This allows you to make blocks zoom into position. Zoom animation is
pretty basic, it just draws a block, then deletes it and then draws
another block next to the previous blocks position.

It repeats this for as many times as you want, in any dimension(x,y,z).

Note that any zooming blocks will delete anything that is in there path